requirements.txt
Django==1.11.6
mysqlclient==1.3.13
djangorestframework==3.7.1
djangorestframework-xml==1.3.0
elasticsearch==5.0.0
certifi==2017.7.27.1
elasticsearch-dsl==5.0.0
Fabric==1.14.0
celery==4.1.1
uWSGI==2.0.15
Fabric==1.14.0
coreapi==2.3.3
Pygments==2.2.0
Markdown==2.6.9
django-rest-swagger==2.1.2
PyYAML==3.12
django-nose==1.4.5
nose-exclude==0.5.0
coverage==4.4.2
django-celery-results==1.0.1
Pillow==4.3.0
redis==2.10.6
django-celery-beat==1.1.0
celery-once==1.2.0
imgkit==0.1.8
django-extensions==1.9.9
ipython==5.5.0
django-mass-edit==3.1
pysftp==0.2.9
paramiko==2.4.0
cryptography==2.1.4
sqlparse==0.2.4
newrelic==3.2.0.91
django-autocomplete-light==3.2.10 # must ues this version
<<<<<<< HEAD
django-auditlog==0.4.3
=======
django-auditlog==0.4.3
django-cors-headers==2.4.0
>>>>>>> 09909669f33c3392509e59c85d3972a6afeeaeeb
